Documentation fixes

+++ meteoio/plugins/SNIO.h
@@ -71,10 +71,10 @@ class SNIO : public IOInterface {
 		void convertUnits(MeteoData& meteo);
 		void convertUnitsBack(MeteoData& meteo);
 		double cloudiness_to_ilwr (const double& RH, const double& TA, const double& cloudiness );
-		void parseMeteoLine(const std::vector<std::string>& vecLine, const std::string& filepos, 
+		void parseMeteoLine(const std::vector<std::string>& vecLine, const std::string& filepos,
 						const Date& dateStart, const Date& dateEnd, MeteoData& md);
 		bool readStationMetaData(const std::string& metafile, const std::string& stationname, StationData& sd);
-		void readMetaData(unsigned int& nrOfStations);
+		void readMetaData(size_t& nrOfStations);
 		void parseMetaDataLine(const std::vector<std::string>& vecLine, StationData& sd);
 		void cleanup() throw();
